pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.13"
dependencies = [
    "databricks-sdk",
    "langchain-community",
    "langchain-core",
    "langchain-huggingface",
    "langchain-openai",
    "langchain-text-splitters",
    "langchain",
    "langgraph",
    "databricks-vectorsearch",
    "python-dotenv",
    "databricks-langchain>=0.14.0",
]
